+++ ChangeLog.md
@@ -1,5 +1,6 @@
 # constrained-toolbox releases
 
 ## 0.1 (2026-07-03)
-- initial release with --delete and --project options
+- initial release with `--delete` and `--project` options
 - defaults to :z shared bind mounts
+- `--caps` to list defined capabilities